Refactor populateUpdate methods
- Changed to method.
- Moved requestIntent to separate file.
